Added Auth page
It is now possible to use signup and signin api calls and authenticate a user through a new auth web page. The forms will add new user to database or check to see if the signin credentials are correct. Upon successful authentication the user will be redirected to the todo list page where their own todos (from the database) are shown.